Revert "fixed the bugs"
