+++ config/config.py
@@ -1,0 +1,34 @@
+import os
+from decouple import config
+from datetime import timedelta
+
+
+BASE_DIR=os.path.dirname(os.path.realpath(__file__))
+
+
+class config:
+    SECRET_KEY=config('SECRET_KEY', 'secret')
+    JWT_ACCESS_TOKEN_EXPIRES=timedelta(minutes=30)
+    JWT_REFRESH_TOKEN_EXPIRES=timedelta(minutes=30)
+    JWT_SECRET_KEY=config('JWT_SECRET_KEY')
+
+
+class DevConfig(config):
+    DEBUG=config()
+    SQLALCHEMY_ECHO=True
+    SQLALCHEMY_DATABASE_URI='sqlite:///'+os.path.join(BASE_DIR,'db.sqlite3')
+    SQLALCHEMY_TRACK_MODIFICATIONS=False
+
+    
+
+class TestConfig(config):
+    pass
+
+class ProdConfig(config):
+    pass
+
+config_dict={
+    'dev':DevConfig,
+    'prod':ProdConfig,
+    'test':TestConfig
+}
