+++ runserver.py
@@ -4,5 +4,6 @@ from api import create_app
 app = create_app()
 
 
-if __name__ == "__main__":
-    app.run(debug=True)
+if __name__ == "__main":
+    app.run(Debug=True)
+
