+++ utils/__init__.py
@@ -1,0 +1,4 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db=SQLAlchemy()
+
